Update base URLs and labels
